pyproject.toml
[tool.poetry]
name = "trailblazer"
version = "21.10.7"
description = "Keep track of and manage analyses"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
include = [
  "README.md",
  "trailblazer/**/*",
  "tests/**/*"
]


[tool.poetry.dependencies]
"ruamel.yaml" = "*"
dependency-injector = "^4.45.0"
Flask = "*"
Flask-Cors = "*"
MarkupSafe = "*"
PyMySQL = "*"
SQLAlchemy = "*"
alembic = "*"
click = "*"
coloredlogs = "*"
cryptography = "*"
flask-reverse-proxy = "*"
gunicorn = "*"
marshmallow = "*"
pydantic = "*"
python = "^3.12"
requests = "*"
setuptools  = "*"
tabulate = "*"
python-keycloak = ">=5.1.2,<6.0.0"

[tool.poetry.group.dev.dependencies]

coveralls =  "^3"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
requests-mock = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
trailblazer = "trailblazer.cli.core:base"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
include_trailing_comma = "true"
line_length=100
multi_line_output=3
ensure_newline_before_comments = "true"

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["alembic"]
ignore = ["E501"]
